requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3

# Development and testing dependencies
pytest==7.4.3
flake8==6.1.0
black==23.11.0
